requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-jwt-extended>=4.6.0
flask-cors>=4.0.0
werkzeug>=3.0.1
requests>=2.31.0
apscheduler>=3.10.4
pymysql>=1.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
face_recognition>=1.3.0
opencv-python>=4.9.0
numpy>=1.26.0
twilio>=8.10.0
fpdf2>=2.7.9
reportlab>=4.1.0
pandas>=2.1.0
openpyxl>=3.1.0
qrcode>=7.4.2
